GPUGraphs

Documentation for GPUGraphs.

GPUGraphs.AbstractSparseGPUMatrix

GPUGraphs.SparseGPUMatrixCSR

GPUGraphs.SparseGPUVector

GPUGraphs.monoid_neutral

GPUGraphs.AbstractSparseGPUMatrix — Type

Abstract type for a GPU compatible sparse matrix

GPUGraphs.SparseGPUMatrixCSR — Type

SparseGPUMatrixCSR{Tv,Ti<:Integer} <: SparseArrays.AbstractSparseMatrix{Tv,Ti}
Tv : Type of the stored values
Ti : Type of the stored indices
A sparse matrix in Compressed Sparse Row format for GPU graph processing.

The matrix is stored in CSR format, with the following fields:
- `m::Int` : Number of rows
- `n::Int` : Number of columns
- `rowptr::AbstractGPUVector{Ti}` : Row i is in rowptr[i]:(rowptr[i+1]-1)
- `colval::AbstractGPUVector{Ti}` : Col indices of stored values
- `nzval::AbstractGPUVector{Tv}` : Stored values, typically nonzeros

GPUGraphs.SparseGPUVector — Type

SparseGPUVector{Tv,Ti<:Integer} <: AbstractSparseVector{Tv,Ti}
Tv : Type of the stored values
Ti : Type of the stored indices
A sparse vector in Compressed Sparse Row format.

GPUGraphs.monoid_neutral — Method

monoid_neutral(T, op)

Return the neutral element of the operation op for the type T.